refactor and merge related updates
    - rebase on master
    - move ES query helpers classes out of search service impl
    - add check for fromRecord
    - addditional junit test for fromRecord
    - append terminology criteria to ES query when relevant
