+++ src/main/java/gov/nih/nci/evs/api/configuration/TerminologyCacheLoader.java
@@ -73,9 +73,7 @@ public class TerminologyCacheLoader implements ApplicationListener<ApplicationRe
               log.info("  get synonym sources ");
               sparqlQueryManagerService.getSynonymSources(terminology);
               log.info("    done synonym sources ");
-//              log.info("  get qualifiers ");
-//              sparqlQueryManagerService.getAllQualifiers(terminology, new IncludeParam("minimal"));
-
+              
             } catch (IOException e) {
               log.error("Unexpected error caching = " + terminology, e);
             }
