+++ src/main/java/gov/nih/nci/evs/api/service/ElasticSearchServiceImpl.java
@@ -2,14 +2,10 @@ package gov.nih.nci.evs.api.service;
 
 import java.io.IOException;
 import java.util.ArrayList;
-import java.util.Collections;
 import java.util.List;
-import java.util.Map;
 
 import org.apache.commons.lang3.StringUtils;
 import org.apache.lucene.search.join.ScoreMode;
-import org.elasticsearch.action.search.SearchResponse;
-import org.elasticsearch.common.text.Text;
 import org.elasticsearch.common.unit.Fuzziness;
 import org.elasticsearch.index.query.BoolQueryBuilder;
 import org.elasticsearch.index.query.MultiMatchQueryBuilder.Type;
@@ -17,32 +13,23 @@ import org.elasticsearch.index.query.Operator;
 import org.elasticsearch.index.query.QueryBuilder;
 import org.elasticsearch.index.query.QueryBuilders;
 import org.elasticsearch.index.query.QueryStringQueryBuilder;
-import org.elasticsearch.search.SearchHit;
 import org.elasticsearch.search.fetch.subphase.highlight.HighlightBuilder;
-import org.elasticsearch.search.fetch.subphase.highlight.HighlightField;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Page;
-import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Pageable;
-import org.springframework.data.domain.Sort;
 import org.springframework.data.elasticsearch.core.ElasticsearchOperations;
-import org.springframework.data.elasticsearch.core.SearchResultMapper;
-import org.springframework.data.elasticsearch.core.aggregation.AggregatedPage;
-import org.springframework.data.elasticsearch.core.aggregation.impl.AggregatedPageImpl;
 import org.springframework.data.elasticsearch.core.query.NativeSearchQueryBuilder;
-import org.springframework.data.elasticsearch.core.query.SearchQuery;
 import org.springframework.stereotype.Service;
 import org.springframework.util.CollectionUtils;
 import org.springframework.web.client.HttpClientErrorException;
 
-import com.fasterxml.jackson.core.JsonProcessingException;
-import com.fasterxml.jackson.databind.ObjectMapper;
-
 import gov.nih.nci.evs.api.model.Concept;
 import gov.nih.nci.evs.api.model.ConceptResultList;
-import gov.nih.nci.evs.api.support.SearchCriteria;
+import gov.nih.nci.evs.api.model.SearchCriteria;
+import gov.nih.nci.evs.api.support.es.EVSConceptResultMapper;
+import gov.nih.nci.evs.api.support.es.EVSPageable;
 
 @Service
 public class ElasticSearchServiceImpl implements ElasticSearchService {
@@ -94,14 +81,21 @@ public class ElasticSearchServiceImpl implements ElasticSearchService {
     .should(QueryBuilders.nestedQuery("definitions", queryStringQueryBuilder, ScoreMode.Total).boost(4f))
     .should(QueryBuilders.nestedQuery("synonyms", queryStringQueryBuilder, ScoreMode.Total).boost(8f));
     
+    //append terminology query
+    QueryBuilder terminologyQuery = buildTerminologyQuery(searchCriteria);
+    if (terminologyQuery != null) {
+      boolQuery = boolQuery.must(terminologyQuery);
+    }
+    
+    //append criteria queries
     List<QueryBuilder> criteriaQueries = buildCriteriaQueries(searchCriteria);
     if (!CollectionUtils.isEmpty(criteriaQueries)) {
-      for(QueryBuilder query: criteriaQueries) {
-        boolQuery = boolQuery.must(query);
+      for(QueryBuilder criteriaQuery: criteriaQueries) {
+        boolQuery = boolQuery.must(criteriaQuery);
       }
     }
     
-    //search query
+    //build final search query
     NativeSearchQueryBuilder searchQuery = new NativeSearchQueryBuilder()
       .withQuery(boolQuery)
       .withIndices(ElasticOperationsService.CONCEPT_INDEX)
@@ -173,6 +167,24 @@ public class ElasticSearchServiceImpl implements ElasticSearchService {
     }
     
     return builder.toString();
+  }
+  
+  /**
+   * builds terminology query based on input search criteria 
+   * 
+   * @param searchCriteria
+   * @return the terminology query builder
+   */
+  private QueryBuilder buildTerminologyQuery(SearchCriteria searchCriteria) {
+    if (CollectionUtils.isEmpty(searchCriteria.getTerminology())) return null;
+    
+    BoolQueryBuilder terminologyQuery = QueryBuilders.boolQuery();
+    
+    for(String terminology: searchCriteria.getTerminology()) {
+      terminologyQuery = terminologyQuery.should(QueryBuilders.matchQuery("terminology", terminology));
+    }
+    
+    return terminologyQuery;
   }
   
   /**
@@ -349,90 +361,5 @@ public class ElasticSearchServiceImpl implements ElasticSearchService {
     
     //nested query on properties
     return QueryBuilders.nestedQuery("synonyms", fieldBoolQuery, ScoreMode.Total);
-  }
-  
-  /**
-   * Custom pageable implementation to support arbitrary offset (work in progress)
-   * 
-   * @author Arun
-   *
-   */
-  private class EVSPageable extends PageRequest {
-    private int offset;
-    public EVSPageable(int page, int size, int offset){
-      super(page, size, Sort.unsorted());
-      this.offset=offset;
-    }
-    
-    @Override
-    public long getOffset(){
-      logger.info("EVSPageable::getOffset() - {}", offset);
-      return this.offset;
-    }
-    
-    //this is a hack to work around spring data bug
-    @Override
-    public int getPageNumber() {
-      return offset / super.getPageSize();
-    }
-
-    @Override
-    public int getPageSize() {
-      return super.getPageSize();
-    }
-  }
-  
-  /**
-   * Custom concept result mapper to extract highlights from search hits
-   * 
-   * @author Arun
-   *
-   */
-  private class EVSConceptResultMapper implements SearchResultMapper {
-    
-    private ObjectMapper mapper;
-    
-    public EVSConceptResultMapper() {
-      this.mapper = new ObjectMapper();
-    }
-    
-    @Override
-    public <T> AggregatedPage<T> mapResults(SearchResponse response, Class<T> clazz, Pageable pageable) {
-      if (response.getHits().getHits().length <= 0) {
-        return new AggregatedPageImpl(Collections.emptyList(), pageable, 0L);
-      }
-      
-      List<Concept> content = new ArrayList<Concept>();
-      for (SearchHit searchHit : response.getHits()) {
-        Concept concept = (Concept) mapSearchHit(searchHit, clazz);
-        content.add(concept);
-      }
-
-      return new AggregatedPageImpl((List<T>) content, pageable, 
-          response.getHits().getTotalHits(), response.getAggregations(), 
-          response.getScrollId(), response.getHits().getMaxScore());
-    }
-
-    @Override
-    public <T> T mapSearchHit(SearchHit searchHit, Class<T> clazz) {
-      Concept concept = null;
-      try {
-        concept = mapper.readValue(searchHit.getSourceAsString(), Concept.class);
-
-        Map<String, HighlightField> highlightMap = searchHit.getHighlightFields();
-        for (String key : highlightMap.keySet()) {
-          HighlightField field = highlightMap.get(key);
-          for (Text text : field.getFragments()) {
-            String highlight = text.string();
-            concept.getHighlights().put(highlight.replaceAll("<em>", "").replaceAll("</em>", ""), highlight);
-          }
-        }
-
-      } catch (JsonProcessingException e) {
-        logger.error(e.getMessage(), e);
-      }
-      
-      return (T) concept;
-    }
   }
 }
